chapter-11/02-data-value/02-streaming-io-sidecar-apache-spark-structured-streaming-heartbeat/requirements.txt
delta-spark==4.0.0
pyspark==4.0.0
